+++ src/extension.ts
@@ -484,25 +484,24 @@ class ExtensionManager implements vscode.Disposable {
         }
 
         const optsVars: KitContextVars = {
+            // sourceDirectory cannot be defined based on any of the below variables.
+            buildKit: '${buildKit}',
+            buildType: '${buildType}',
+            buildKitVendor: '${buildKitVendor}',
+            buildKitTriple: '${buildKitTriple}',
+            buildKitVersion: '${buildKitVersion}',
+            buildKitHostOs: '${buildKitVendor}',
+            buildKitTargetOs: '${buildKitTargetOs}',
+            buildKitTargetArch: '${buildKitTargetArch}',
+            buildKitVersionMajor: '${buildKitVersionMajor}',
+            buildKitVersionMinor: '${buildKitVersionMinor}',
+            generator: '${generator}',
             userHome: paths.userHome,
             workspaceFolder: cmakeProject.workspaceContext.folder.uri.fsPath,
             workspaceFolderBasename: cmakeProject.workspaceContext.folder.name,
+            workspaceHash: '${workspaceHash}',
             workspaceRoot: cmakeProject.workspaceContext.folder.uri.fsPath,
-            workspaceRootFolderName: cmakeProject.workspaceContext.folder.name,
-
-            // sourceDirectory cannot be defined based on any of the below variables.
-            buildKit: "",
-            buildType: "",
-            generator: "",
-            buildKitVendor: "",
-            buildKitTriple: "",
-            buildKitVersion: "",
-            buildKitHostOs: "",
-            buildKitTargetOs: "",
-            buildKitTargetArch: "",
-            buildKitVersionMajor: "",
-            buildKitVersionMinor: "",
-            workspaceHash: ""
+            workspaceRootFolderName: cmakeProject.workspaceContext.folder.name
         };
 
         const sourceDirectory: string = cmakeProject.workspaceContext.config.sourceDirectory;