+++ src/preset.ts
@@ -368,14 +368,6 @@ export function setOriginalUserPresetsFile(folder: string, presets: PresetsFile 
     originalUserPresetsFiles.set(folder, presets);
 }
 
-export function getPresetsFile(folder: string) {
-    return presetsFiles.get(folder);
-}
-
-export function getUserPresetsFile(folder: string) {
-    return userPresetsFiles.get(folder);
-}
-
 export function setPresetsFile(folder: string, presets: PresetsFile | undefined) {
     presetsFiles.set(folder, presets);
 }
@@ -601,7 +593,7 @@ async function getExpansionOptions(workspaceFolder: string, sourceDir: string, p
     return expansionOpts;
 }
 
-async function expandCondition(condition: boolean | Condition | null | undefined, expansionOpts: ExpansionOptions) {
+async function expandCondition(condition: boolean | Condition | null | undefined, expansionOpts: ExpansionOptions): Promise<boolean | Condition | undefined> {
     if (util.isNullOrUndefined(condition)) {
         return undefined;
     }
@@ -647,21 +639,21 @@ async function expandCondition(condition: boolean | Condition | null | undefined
 }
 
 export async function expandConditionsForPresets(folder: string, sourceDir: string) {
-    for (const preset of configurePresets(folder)) {
-        const opts = await getExpansionOptions('${workspaceFolder}', sourceDir, preset);
+    for (const preset of allConfigurePresets(folder)) {
         if (preset.condition) {
+            const opts = await getExpansionOptions('${workspaceFolder}', sourceDir, preset);
             preset.condition = await expandCondition(preset.condition, opts);
         }
     }
-    for (const preset of buildPresets(folder)) {
-        const opts = await getExpansionOptions('${workspaceFolder}', sourceDir, preset);
+    for (const preset of allBuildPresets(folder)) {
         if (preset.condition) {
+            const opts = await getExpansionOptions('${workspaceFolder}', sourceDir, preset);
             preset.condition = await expandCondition(preset.condition, opts);
         }
     }
-    for (const preset of testPresets(folder)) {
-        const opts = await getExpansionOptions('${workspaceFolder}', sourceDir, preset);
+    for (const preset of allTestPresets(folder)) {
         if (preset.condition) {
+            const opts = await getExpansionOptions('${workspaceFolder}', sourceDir, preset);
             preset.condition = await expandCondition(preset.condition, opts);
         }
     }